Refactor result module: reorganize imports and move scene-related files to visualizer
