+++ model_api/python/model_api/models/result/anomaly.py
