+++ model_api/python/model_api/models/result/visual_prompting.py
