+++ model_api/python/model_api/visualizer/scene/anomaly.py
@@ -5,10 +5,11 @@
 
 from PIL import Image
 
-from model_api.models.result.types import AnomalyResult
+from model_api.models.result import AnomalyResult
 from model_api.visualizer.layout import Flatten, Layout
 from model_api.visualizer.primitive import Overlay
-from model_api.visualizer.scene import Scene
+
+from .scene import Scene
 
 
 class AnomalyScene(Scene):
