+++ model_api/python/model_api/visualizer/scene/keypoint.py
@@ -3,10 +3,11 @@
 # Copyright (C) 2024 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 
-from model_api.models.result.types import DetectedKeypoints
+from model_api.models.result import DetectedKeypoints
 from model_api.visualizer.layout import Flatten, Layout
 from model_api.visualizer.primitive import Overlay
-from model_api.visualizer.scene import Scene
+
+from .scene import Scene
 
 
 class KeypointScene(Scene):
