+++ model_api/python/model_api/visualizer/scene/segmentation.py
@@ -3,8 +3,9 @@
 # Copyright (C) 2024 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 
-from model_api.models.result.types import InstanceSegmentationResult
-from model_api.visualizer.scene import Scene
+from model_api.models.result import InstanceSegmentationResult
+
+from .scene import Scene
 
 
 class SegmentationScene(Scene):
